typeconversion

+++ src/main/java/typeconversion/whichtype/Type.java
@@ -2,7 +2,7 @@ package typeconversion.whichtype;
 
 public enum Type {
 
-    BYTE(Byte.MIN_VALUE, Byte.MAX_VALUE), SHORT(Short.MIN_VALUE, Short.MAX_VALUE), INT(Long.MIN_VALUE, Long.MAX_VALUE);
+    BYTE(Byte.MIN_VALUE, Byte.MAX_VALUE), SHORT(Short.MIN_VALUE, Short.MAX_VALUE), INT(Integer.MIN_VALUE, Integer.MAX_VALUE);
 
     private final long minValue;
     private final long maxValue;
